# [New] Archlinux install instructions