--- a/install/installation-archlinux.md
+++ b/install/installation-archlinux.md
@@ -116,7 +116,7 @@ out the [TimescaleDB][tsdb-docs] section in our documentation, and find out what
 you can do with it.
 
 If you want to work through some tutorials to help you get up and running with
-TimescaleDB and time-series data, check out our [tutorials][tutorials] section.
+TimescaleDB and time-series data, check out the [tutorials][tutorials] section.
 
 You can always [contact us][contact] if you need help working something out, or
 if you want to have a chat.