--- a/install/installation-archlinux.md
+++ b/install/installation-archlinux.md
@@ -61,7 +61,7 @@ information, see the [configuration][config] section.
 When you have PostgreSQL and TimescaleDB installed, you can connect to it from
 your local system using the `psql` command-line utility. This is the same tool
 you might have used to connect to PostgreSQL before, but if you haven't
-installed it yet, check out our [installing psql][install-psql] section.
+installed it yet, check out the [installing psql][install-psql] section.
 
 <procedure>