--- a/install/installation-archlinux.md
+++ b/install/installation-archlinux.md
@@ -112,7 +112,7 @@ tsdb=# \dx
 
 ## Where to next
 Now that you have your first TimescaleDB database up and running, you can check
-out the [TimescaleDB][tsdb-docs] section in our documentation, and find out what
+out the [TimescaleDB][tsdb-docs] section in the documentation, and find out what
 you can do with it.
 
 If you want to work through some tutorials to help you get up and running with