--- a/install/installation-archlinux.md
+++ b/install/installation-archlinux.md
@@ -45,7 +45,7 @@ instead.
     systemctl enable postgresql.service
     systemctl start postgresql.service
     ```
-1.  Restart the service after enabling TimescaleDB with `timescaledb-tune`:
+1.  Restart the service:
     ```bash
     systemctl restart postgresql
     ```