--- a/install/installation-archlinux.md
+++ b/install/installation-archlinux.md
@@ -25,7 +25,7 @@ instead.
     ```
 1.  Initialize the database:
     ```bash
-    initdb --locale=en_US.UTF-8 --encoding=UTF8 -D /var/lib/postgres/data --data-checksums
+    sudo -u postgres initdb --locale=en_US.UTF-8 --encoding=UTF8 -D /var/lib/postgres/data --data-checksums
     ```
 1. Log out the `postgres` user:
    ```bash