--- a/install/page-index/page-index.js
+++ b/install/page-index/page-index.js
@@ -49,6 +49,21 @@ module.exports = [
               "Install self-hosted TimescaleDB on Red Hat or CentOS using dnf",
           },
           {
+            title: "Arch Linux",
+            href: "installation-archlinux",
+            tags: ["install", "archlinux", "timescaledb"],
+            keywords: [
+              "TimescaleDB",
+              "install",
+              "self-hosted",
+              "Arch Linux",
+              "TimescaleDB extension"
+            ],
+            iconSrc: "//assets.iobeam.com/images/docs/redhat_centos.png",
+            excerpt:
+              "Install self-hosted TimescaleDB on Red Hat or CentOS using dnf",
+          },
+          {
             title: "Windows",
             href: "installation-windows",
             tags: ["install", "mswin", "exe", "timescaledb"],