@Rajakavitha1 2022-08-03
> Looking good, please retest the steps before committing in case we missed something.

Just validated the steps:[rajie@localhost ~]$ sudo pacman -Syu timescaledb timescaledb-tune
```
[sudo] password for rajie: 
:: Synchronizing package databases...
 core is up to date
 extra is up to date
 community is up to date
:: Starting full system upgrade...
resolving dependencies...
looking for conflicting packages...

Packages (6) libxslt-1.1.35-1  llvm-libs-14.0.6-2  postgresql-14.3-3
             postgresql-libs-14.3-3  timescaledb-2.7.2-1
             timescaledb-tune-0.13.0-1

Total Download Size:    49.28 MiB
Total Installed Size:  191.38 MiB

:: Proceed with installation? [Y/n] Y
:: Retrieving packages...
 llvm-libs-14.0.6...    27.5 MiB  11.2 MiB/s 00:02 [######################] 100%
 postgresql-14.3-...    18.3 MiB  15.6 MiB/s 00:01 [######################] 100%
 postgresql-libs-...  1442.9 KiB  2012 KiB/s 00:01 [######################] 100%
 timescaledb-2.7....  1041.2 KiB  2.44 MiB/s 00:00 [######################] 100%
 timescaledb-tune...   743.1 KiB  1782 KiB/s 00:00 [######################] 100%
 libxslt-1.1.35-1...   401.5 KiB  1004 KiB/s 00:00 [######################] 100%
 Total (6/6)            49.3 MiB  7.75 MiB/s 00:06 [######################] 100%
(6/6) checking keys in keyring                     [######################] 100%
(6/6) checking package integrity                   [######################] 100%
(6/6) loading package files                        [######################] 100%
(6/6) checking for file conflicts                  [######################] 100%
(6/6) checking available disk space                [######################] 100%
:: Processing package changes...
(1/6) installing postgresql-libs                   [######################] 100%
(2/6) installing llvm-libs                         [######################] 100%
(3/6) installing libxslt                           [######################] 100%
Optional dependencies for libxslt
    python: Python bindings [installed]
(4/6) installing postgresql                        [######################] 100%
Optional dependencies for postgresql
    python: for PL/Python 3 support [installed]
    perl: for PL/Perl support [installed]
    tcl: for PL/Tcl support
    postgresql-old-upgrade: upgrade from previous major version using pg_upgrade
    logrotate: rotates system logs automatically
(5/6) installing timescaledb                       [######################] 100%
To use TimescaleDB as PostgreSQL extension,
put in /var/lib/postgres/data/postgresql.conf:
'shared_preload_libraries = "timescaledb"'
If multiple libraries are required, separate them with comma.
Then restart postgresql:
$ systemctl restart postgresql
Optional dependencies for timescaledb
    timescaledb-tune: tune postgresql.conf for better performance [pending]
    timescaledb-old-upgrade: upgrade from previous major version of PostgreSQL
(6/6) installing timescaledb-tune                  [######################] 100%
:: Running post-transaction hooks...
(1/4) Creating system user accounts...
Creating group 'postgres' with GID 974.
Creating user 'postgres' (PostgreSQL user) with UID 974 and GID 974.
(2/4) Reloading system manager configuration...
(3/4) Creating temporary files...
[rajie@localhost ~]$ sudo -u postgres initdb --locale=en_US.UTF-8 --encoding=UTF8 -D /var/lib/postgres/data --data-checksums
could not change directory to "/home/rajie": Permission denied
The files belonging to this database system will be owned by user "postgres".
This user must also own the server process.

The database cluster will be initialized with locale "en_US.UTF-8".
The default text search configuration will be set to "english".

Data page checksums are enabled.

fixing permissions on existing directory /var/lib/postgres/data ... ok
creating subdirectories ... ok
selecting dynamic shared memory implementation ... posix
selecting default max_connections ... 100
selecting default shared_buffers ... 128MB
selecting default time zone ... UTC
creating configuration files ... ok
running bootstrap script ... ok
performing post-bootstrap initialization ... ok
syncing data to disk ... ok

initdb: warning: enabling "trust" authentication for local connections
You can change this by editing pg_hba.conf or using the option -A, or
--auth-local and --auth-host, the next time you run initdb.

Success. You can now start the database server using:

    pg_ctl -D /var/lib/postgres/data -l logfile start

[rajie@localhost ~]$ timescaledb-tune
exit: could not find postgresql.conf at any of these locations:
/etc/postgresql/14/main/postgresql.conf
/var/lib/pgsql/14/data/postgresql.conf
/var/lib/postgres/data/postgresql.conf
/var/lib/postgresql/data/postgresql.conf
[rajie@localhost ~]$ pg_ctl -D /var/lib/postgres/data -l logfile start
pg_ctl: could not open PID file "/var/lib/postgres/data/postmaster.pid": Permission denied
[rajie@localhost ~]$ sudo pg_ctl -D /var/lib/postgres/data -l logfile start
pg_ctl: cannot be run as root
Please log in (using, e.g., "su") as the (unprivileged) user that will
own the server process.
[rajie@localhost ~]$ su
Password: 
su: Authentication failure
[rajie@localhost ~]$ su
Password: 
[root@localhost rajie]# pg_ctl -D /var/lib/postgres/data -l logfile start
pg_ctl: cannot be run as root
Please log in (using, e.g., "su") as the (unprivileged) user that will
own the server process.
[root@localhost rajie]# su rajie
[rajie@localhost ~]$ timescaledb-tune
exit: could not find postgresql.conf at any of these locations:
/etc/postgresql/14/main/postgresql.conf
/var/lib/pgsql/14/data/postgresql.conf
/var/lib/postgres/data/postgresql.conf
/var/lib/postgresql/data/postgresql.conf
[rajie@localhost ~]$ cd /var/lib
[rajie@localhost lib]$ ls
arpd   dbus      krb5kdc   misc    portables  private  tpm2-tss
cloud  dhclient  machines  pacman  postgres   systemd
[rajie@localhost lib]$ cd postgres
[rajie@localhost postgres]$ ls
data
[rajie@localhost postgres]$ cd data
bash: cd: data: Permission denied
[rajie@localhost postgres]$ sudo timescaledb-tune
Using postgresql.conf at this path:
/var/lib/postgres/data/postgresql.conf

Is this correct? [(y)es/(n)o]: y
Writing backup to:
/tmp/timescaledb_tune.backup202208031335

shared_preload_libraries needs to be updated
Current:
#shared_preload_libraries = ''
Recommended:
shared_preload_libraries = 'timescaledb'
Is this okay? [(y)es/(n)o]: y
success: shared_preload_libraries will be updated

Tune memory/parallelism/WAL and other settings? [(y)es/(n)o]: y
Recommendations based on 3.83 GB of available memory and 2 CPUs for PostgreSQL 14

Memory settings recommendations
Current:
shared_buffers = 128MB
#effective_cache_size = 4GB
#maintenance_work_mem = 64MB
#work_mem = 4MB
Recommended:
shared_buffers = 1004692kB
effective_cache_size = 2943MB
maintenance_work_mem = 502346kB
work_mem = 10046kB
Is this okay? [(y)es/(s)kip/(q)uit]: y
success: memory settings will be updated

Parallelism settings recommendations
Current:
missing: timescaledb.max_background_workers
#max_worker_processes = 8
#max_parallel_workers_per_gather = 2
#max_parallel_workers = 8
Recommended:
timescaledb.max_background_workers = 8
max_worker_processes = 13
max_parallel_workers_per_gather = 1
max_parallel_workers = 2
Is this okay? [(y)es/(s)kip/(q)uit]: y
success: parallelism settings will be updated

WAL settings recommendations
Current:
#wal_buffers = -1
min_wal_size = 80MB
Recommended:
wal_buffers = 16MB
min_wal_size = 512MB
Is this okay? [(y)es/(s)kip/(q)uit]: y
success: WAL settings will be updated

Miscellaneous settings recommendations
Current:
#default_statistics_target = 100
#random_page_cost = 4.0
#checkpoint_completion_target = 0.9
max_connections = 100
#max_locks_per_transaction = 64
#autovacuum_max_workers = 3
#autovacuum_naptime = 1min
#effective_io_concurrency = 1
Recommended:
default_statistics_target = 500
random_page_cost = 1.1
checkpoint_completion_target = 0.9
max_connections = 50
max_locks_per_transaction = 64
autovacuum_max_workers = 10
autovacuum_naptime = 10
effective_io_concurrency = 256
Is this okay? [(y)es/(s)kip/(q)uit]: y
success: miscellaneous settings will be updated
Saving changes to: /var/lib/postgres/data/postgresql.conf
[rajie@localhost postgres]$ systemctl enable postgresql.service
Failed to enable unit: Access denied
[rajie@localhost postgres]$ sudo systemctl enable postgresql.service
Created symlink /etc/systemd/system/multi-user.target.wants/postgresql.service → /usr/lib/systemd/system/postgresql.service.
[rajie@localhost postgres]$ sudo systemctl start postgresql.service
[rajie@localhost postgres]$ sudo systemctl status
● localhost
    State: running
    Units: 234 loaded (incl. loaded aliases)
     Jobs: 0 queued
   Failed: 0 units
    Since: Wed 2022-08-03 13:02:18 UTC; 34min ago
  systemd: 251.3-1-arch
   CGroup: /
           ├─init.scope
           │ └─1 /usr/lib/systemd/systemd --system --deserialize 23
           ├─system.slice
           │ ├─dbus.service
           │ │ └─302 /usr/bin/dbus-daemon --system --address=systemd: --nofork --nopidfile>
           │ ├─postgresql.service
           │ │ ├─5206 /usr/bin/postgres -D /var/lib/postgres/data
           │ │ ├─5209 "postgres: checkpointer "
           │ │ ├─5210 "postgres: background writer "
           │ │ ├─5211 "postgres: walwriter "
           │ │ ├─5212 "postgres: autovacuum launcher "
           │ │ ├─5213 "postgres: stats collector "
           │ │ ├─5214 "postgres: TimescaleDB Background Worker Launcher "
           │ │ └─5215 "postgres: logical replication launcher "
           │ ├─sshd.service
[rajie@localhost postgres]$ 
[rajie@localhost postgres]$ sudo -u postgres psql
psql (14.3)
Type "help" for help.

postgres=# CREATE database tsdb;
CREATE DATABASE
postgres=# \c tsdb
You are now connected to database "tsdb" as user "postgres".
tsdb=# CREATE EXTENSION IF NOT EXISTS timescaledb;
WARNING:  
WELCOME TO
 _____ _                               _     ____________  
|_   _(_)                             | |    |  _  \ ___ \ 
  | |  _ _ __ ___   ___  ___  ___ __ _| | ___| | | | |_/ / 
  | | | |  _ ` _ \ / _ \/ __|/ __/ _` | |/ _ \ | | | ___ \ 
  | | | | | | | | |  __/\__ \ (_| (_| | |  __/ |/ /| |_/ /
  |_| |_|_| |_| |_|\___||___/\___\__,_|_|\___|___/ \____/
               Running version 2.7.2
For more information on TimescaleDB, please visit the following links:

 1. Getting started: https://docs.timescale.com/timescaledb/latest/getting-started
 2. API reference documentation: https://docs.timescale.com/api/latest
 3. How TimescaleDB is designed: https://docs.timescale.com/timescaledb/latest/overview/core-concepts

Note: TimescaleDB collects anonymous reports to better understand and assist our users.
For more information and how to disable, please see our docs https://docs.timescale.com/timescaledb/latest/how-to-guides/configuration/telemetry.

CREATE EXTENSION
tsdb=# 
```